bump to 2mb.  DO NOT USE WITH BTCWALLET.
Allows blocks that are 2mb or smaller.

Won't check for the 750/1000 block threshold.  Also won't check for the grace period.  SO DO NOT USE WITH BTCWALLET.

In other words, this code will accept blocks that nobody else will accept if it's before the 750/1000 block threshold, or before the 28 day grace period.